Fixes the errors reported by Sid / G3VBV while compiling on a 64 bit platform.
git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@1094 f8843c6b-282e-0410-804b-c0d5895947d2

+++ src/async/audio/AsyncAudioDevice.cpp
@@ -550,7 +550,7 @@ void AudioDevice::writeSpaceAvailable(FdWatch *watch)
     //fragments = max(0, 2 - (info.fragstotal - info.fragments));
     fragments = info.fragments;
     fragsize = info.fragsize / sizeof(short);
-    samples_to_write = min(sizeof(buf) / sizeof(*buf),
+    samples_to_write = min(static_cast<unsigned>(sizeof(buf) / sizeof(*buf)),
     		fragments * fragsize);
     //samples_to_write = min(samples_to_write, 7936U);
